# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "E2BR3 Backend API",
    "description": "OpenAPI documentation for the Axum-based E2BR3 backend. Authenticated API routes use the HTTP-only `auth-token` cookie. Internal machine-to-machine routes use the `x-callback-token` header.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/cases/{case_id}/validation": {
      "get": {
        "tags": [
          "case-subresources"
        ],
        "operationId": "validate_case",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "description": "Case ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "profile",
            "in": "query",
            "description": "Validation profile override",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Case validation report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericDataResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth_token": []
          }
        ]
      }
    },
    "/api/cases/{case_id}/validation/all": {
      "get": {
        "tags": [
          "case-subresources"
        ],
        "operationId": "validate_case_all",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "description": "Case ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All profile validation report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericDataResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth_token": []
          }
        ]
      }
    },
...
    "/api/cases/{id}": {
      "get": {
        "tags": [
          "cases"
        ],
        "operationId": "get_case",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Case ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Case details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            }
          },
          "404": {
            "description": "Case not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth_token": []
          }
        ]
      },
      "put": {
        "tags": [
          "cases"
        ],
        "operationId": "update_case",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Case ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Case updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid case payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Case not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth_token": []
          }
        ]
      },
      "delete": {
        "tags": [
          "cases"
        ],
        "operationId": "delete_case",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Case ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Case deleted"
          },
          "404": {
            "description": "Case not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth_token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "CaseDoc": {
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "safety_report_id",
          "version",
          "status",
          "workflow_status",
          "workflow_updated_at",
          "qc_state",
          "is_locked",
          "can_act_on_workflow",
          "created_by",
          "dirty_c",
          "dirty_d",
          "dirty_e",
          "dirty_f",
          "dirty_g",
          "dirty_h",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "appendices_json": {
            "type": [
              "string",
              "null"
            ]
          },
          "can_act_on_workflow": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string"
          },
          "created_by": {
            "type": "string"
          },
          "dg_prd_key": {
            "type": [
              "string",
              "null"
            ]
          },
          "dirty_c": {
            "type": "boolean"
          },
          "dirty_d": {
            "type": "boolean"
          },
          "dirty_e": {
            "type": "boolean"
          },
          "dirty_f": {
            "type": "boolean"
          },
          "dirty_g": {
            "type": "boolean"
          },
          "dirty_h": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "is_locked": {
            "type": "boolean"
          },
          "mfds_report_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "organization_id": {
            "type": "string"
          },
          "qc_state": {
            "type": "string"
          },
          "raw_xml": {
            "type": [
              "string",
              "null"
            ]
          },
          "report_year": {
            "type": [
              "string",
              "null"
            ]
          },
          "review_receivers_json": {
            "type": [
              "string",
              "null"
            ]
          },
          "safety_report_id": {
            "type": "string"
          },
          "source_document_base64": {
            "type": [
              "string",
              "null"
            ]
          },
          "source_document_media_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "source_document_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "submitted_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "submitted_by": {
            "type": [
              "string",
              "null"
            ]
          },
          "updated_at": {
            "type": "string"
          },
          "updated_by": {
            "type": [
              "string",
              "null"
            ]
          },
          "validation_profile": {
            "type": [
              "string",
              "null"
            ]
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "workflow_assigned_role": {
            "type": [
              "string",
              "null"
            ]
          },
          "workflow_assigned_user_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "workflow_block_reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "workflow_description": {
            "type": [
              "string",
              "null"
            ]
          },
          "workflow_due_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "workflow_routes_json": {
            "type": [
              "string",
              "null"
            ]
          },
          "workflow_status": {
            "type": "string"
          },
          "workflow_updated_at": {
            "type": "string"
          }
        }
      },
...
      "CaseForUpdateDoc": {
        "type": "object",
        "properties": {
          "appendices_json": {
            "type": [
              "string",
              "null"
            ]
          },
          "dg_prd_key": {
            "type": [
              "string",
              "null"
            ]
          },
          "mfds_report_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "report_year": {
            "type": [
              "string",
              "null"
            ]
          },
          "review_receivers_json": {
            "type": [
              "string",
              "null"
            ]
          },
          "safety_report_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "source_document_base64": {
            "type": [
              "string",
              "null"
            ]
          },
          "source_document_media_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "source_document_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": [
              "string",
              "null"
            ]
          },
          "validation_profile": {
            "type": [
              "string",
              "null"
            ]
          },
          "workflow_routes_json": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "CaseResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CaseDoc"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "GenericDataResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object"
          }
        }
      },
...
      "UpdateCaseRequest": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CaseForUpdateDoc"
          }
        }
      },
...
    },
    "securitySchemes": {
      "auth_token": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth-token"
      },
      "callback_token": {
        "type": "apiKey",
        "in": "header",
        "name": "x-callback-token"
      }
    }
  },
...
}